Print an integer using the "{:d}.format" approach.
If the value is an integer, no exception is raised and the value is
printed accordingly. The function returns True in this case.
However, if the value of the wrong type is passed to the function
an exception is raised and handled by returning a False from the
function signifying the inability to print the value.